+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/fragments/GetStartedFragment.kt
@@ -5,13 +5,14 @@
  *
  */
 
-package com.microsoft.device.display.samples.twonote
+package com.microsoft.device.display.samples.twonote.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.microsoft.device.display.samples.twonote.R
 
 class GetStartedFragment : Fragment() {
