+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/fragments/NoteListFragment.kt
@@ -4,7 +4,7 @@
  *
  */
 
-package com.microsoft.device.display.samples.twonote
+package com.microsoft.device.display.samples.twonote.fragments
 
 import Defines.DETAIL_FRAGMENT
 import Defines.LIST_VIEW
@@ -25,11 +25,14 @@ import androidx.fragment.app.Fragment
 import com.google.android.material.appbar.MaterialToolbar
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.textfield.TextInputEditText
-import com.microsoft.device.display.samples.twonote.includes.DataProvider
-import com.microsoft.device.display.samples.twonote.includes.FileSystem
-import com.microsoft.device.display.samples.twonote.structures.DirEntry
-import com.microsoft.device.display.samples.twonote.structures.INode
-import com.microsoft.device.display.samples.twonote.structures.Note
+import com.microsoft.device.display.samples.twonote.MainActivity
+import com.microsoft.device.display.samples.twonote.R
+import com.microsoft.device.display.samples.twonote.models.DirEntry
+import com.microsoft.device.display.samples.twonote.models.INode
+import com.microsoft.device.display.samples.twonote.models.Note
+import com.microsoft.device.display.samples.twonote.utils.DataProvider
+import com.microsoft.device.display.samples.twonote.utils.FileSystem
+import com.microsoft.device.display.samples.twonote.utils.NoteSelectionListener
 import com.microsoft.device.dualscreen.core.ScreenHelper
 
 class NoteListFragment : Fragment(), AdapterView.OnItemClickListener, AdapterView.OnItemLongClickListener, AdapterView.OnItemSelectedListener {
