+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/models/SerializedImage.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.microsoft.device.display.samples.twonote.structures
+package com.microsoft.device.display.samples.twonote.models
 
 import java.io.Serializable
