+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/utils/DragHandler.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.microsoft.device.display.samples.twonote.includes
+package com.microsoft.device.display.samples.twonote.utils
 
 import Defines.IMAGE_PREFIX
 import Defines.TEXT_PREFIX
@@ -15,8 +15,8 @@ import android.view.DragEvent
 import android.widget.ImageView
 import androidx.core.app.ActivityCompat
 import com.microsoft.device.display.samples.twonote.MainActivity
-import com.microsoft.device.display.samples.twonote.NoteDetailFragment
-import com.microsoft.device.display.samples.twonote.structures.SerializedImage
+import com.microsoft.device.display.samples.twonote.fragments.NoteDetailFragment
+import com.microsoft.device.display.samples.twonote.models.SerializedImage
 
 /* Class used to update app contents when an appropriate drag event occurs */
 class DragHandler(private val fragment: NoteDetailFragment) {
