+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/utils/FileSystem.kt
@@ -1,11 +1,11 @@
-package com.microsoft.device.display.samples.twonote.includes
+package com.microsoft.device.display.samples.twonote.utils
 
 import android.content.Context
 import android.util.Log
 import com.microsoft.device.display.samples.twonote.R
-import com.microsoft.device.display.samples.twonote.structures.DirEntry
-import com.microsoft.device.display.samples.twonote.structures.INode
-import com.microsoft.device.display.samples.twonote.structures.Note
+import com.microsoft.device.display.samples.twonote.models.DirEntry
+import com.microsoft.device.display.samples.twonote.models.INode
+import com.microsoft.device.display.samples.twonote.models.Note
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
