+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/utils/ImageHandler.kt
@@ -5,7 +5,7 @@
  *
  */
 
-package com.microsoft.device.display.samples.twonote.includes
+package com.microsoft.device.display.samples.twonote.utils
 
 import Defines.LAND_TO_PORT
 import Defines.MIN_DIMEN
@@ -29,8 +29,8 @@ import android.widget.ImageView
 import android.widget.RelativeLayout
 import androidx.core.view.drawToBitmap
 import com.microsoft.device.display.samples.twonote.MainActivity
-import com.microsoft.device.display.samples.twonote.NoteDetailFragment
-import com.microsoft.device.display.samples.twonote.structures.SerializedImage
+import com.microsoft.device.display.samples.twonote.fragments.NoteDetailFragment
+import com.microsoft.device.display.samples.twonote.models.SerializedImage
 import java.io.ByteArrayOutputStream
 import java.util.Calendar
 import kotlin.math.abs
