+++ TwoNote/app/src/main/java/com/microsoft/device/display/samples/twonote/utils/TextHandler.kt
